bigdecimal

+++ apps/backend/src/models/db.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use chrono::{DateTime, Utc};
 use sqlx::FromRow;
 use uuid::Uuid;
@@ -24,9 +25,9 @@ pub struct MarketRow {
     pub id: String,
     pub base_ticker: String,
     pub quote_ticker: String,
-    pub tick_size: u128,
-    pub lot_size: u128,
-    pub min_size: u64,
+    pub tick_size: BigDecimal,
+    pub lot_size: BigDecimal,
+    pub min_size: BigDecimal,
     pub maker_fee_bps: i32,
     pub taker_fee_bps: i32,
 }
@@ -36,13 +37,13 @@ pub struct OrderRow {
     pub id: Uuid,
     pub user_address: String,
     pub market_id: String,
-    pub price: u128,
-    pub size: u128,
+    pub price: BigDecimal,
+    pub size: BigDecimal,
     pub side: String, // Custom type 'side' in DB, stored as TEXT
     #[sqlx(rename = "type")]
     pub order_type: String, // Custom type 'order_type' in DB
     pub status: String, // Custom type 'order_status' in DB
-    pub filled_size: u128,
+    pub filled_size: BigDecimal,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
@@ -55,8 +56,8 @@ pub struct TradeRow {
     pub seller_address: String,
     pub buyer_order_id: Uuid,
     pub seller_order_id: Uuid,
-    pub price: u128,
-    pub size: u128,
+    pub price: BigDecimal,
+    pub size: BigDecimal,
     pub timestamp: DateTime<Utc>,
 }
 
@@ -64,8 +65,8 @@ pub struct TradeRow {
 pub struct BalanceRow {
     pub user_address: String,
     pub token_ticker: String,
-    pub amount: u128,
-    pub open_interest: u128,
+    pub amount: BigDecimal,
+    pub open_interest: BigDecimal,
     pub updated_at: DateTime<Utc>,
 }
 
@@ -73,9 +74,9 @@ pub struct BalanceRow {
 pub struct CandleRow {
     pub market_id: String,
     pub timestamp: DateTime<Utc>,
-    pub open: u128,
-    pub high: u128,
-    pub low: u128,
-    pub close: u128,
-    pub volume: u128,
+    pub open: BigDecimal,
+    pub high: BigDecimal,
+    pub low: BigDecimal,
+    pub close: BigDecimal,
+    pub volume: BigDecimal,
 }
